Simplify trait bound and implement Transport directly on plain types.
Previously, the `Transport` trait was always implemented on mutable
reference to avoid the need to inject a lifetime in the associated type.
This is now done with a helper trait.

Additionally, this commit moves trait bounds around to simplify them, at
the const of no longer supporting !Send transports (and similar).