Make Encode/DecodeBody traits on Format instead of T.
Someone implementing a new format needs to be able to define the
encode/decode implementations. If the trait is implemented on T, then
only fizyr_rpc or the crate containing T could define the impl.

The down-side is that the crate with the format needs to define all
impls now.