Remove `recv_*_update` functions in favor of `update.as_*()`.
The `recv_*_update()` functions had to put messages back in the peek
buffer when they were the wrong kind of update. However, this is not in
line with the behaviour of `recv_udate()`. So it's complicated niche
behaviour that would be hard to document correctly.

With this change, users will do `req.recv_update()?.into_foo()` instead
of `req.recv_foo_update()?`. This completely sidesteppes the issue of
putting messages back in the peek buffer.