Do not consume request handles when sending a response.
There is no point in that anymore, since you can already create as many
write handles for the request as you want.

However, we need to ensure that `recv` functions do not wait
indefinately when a request is no longer active, so we use a new
RequestHandleCommand channel instead of a raw message channel, which can
be used to instruct the read handle to close the channel.